requirements.txt
SpeechRecognition==3.10.0
pyttsx3==2.90
pywhatkit==5.4
PyAudio==0.2.13
requests==2.31.0
openai==0.28.1
python-dotenv==1.0.0
schedule==1.2.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0

